refactor: separate component logic to hooks
